uk-esos-app-api/src/main/java/uk/gov/esos/api/reporting/noc/phase3/validation/NocP3LeadAssessorContextValidatorService.java
package uk.gov.esos.api.reporting.noc.phase3.validation;

import org.springframework.stereotype.Service;
import uk.gov.esos.api.reporting.noc.common.domain.NocValidationResult;
import uk.gov.esos.api.reporting.noc.common.validation.NocSectionConstraintValidatorService;
import uk.gov.esos.api.reporting.noc.phase3.domain.NocP3Container;
import uk.gov.esos.api.reporting.noc.phase3.domain.ReportingObligationCategory;
import uk.gov.esos.api.reporting.noc.phase3.domain.leadassessor.LeadAssessor;

import java.util.Set;

@Service
public class NocP3LeadAssessorContextValidatorService extends NocP3SectionValidatorService<LeadAssessor> implements NocP3SectionContextValidator {

    public NocP3LeadAssessorContextValidatorService(NocSectionConstraintValidatorService<LeadAssessor> nocSectionConstraintValidatorService) {
        super(nocSectionConstraintValidatorService);
    }

    @Override
    public NocValidationResult validate(NocP3Container nocContainer, ReportingObligationCategory reportingObligationCategory) {
		LeadAssessor section = nocContainer.getNoc().getLeadAssessor();
        return this.validate(section, nocContainer, reportingObligationCategory);
    }

    @Override
    protected Set<ReportingObligationCategory> getApplicableReportingObligationCategories() {
        return Set.of(
                ReportingObligationCategory.ESOS_ENERGY_ASSESSMENTS_95_TO_100,
                ReportingObligationCategory.PARTIAL_ENERGY_ASSESSMENTS,
                ReportingObligationCategory.ALTERNATIVE_ENERGY_ASSESSMENTS_95_TO_100
        );
    }

    @Override
    protected String getSectionName() {
        return LeadAssessor.class.getName();
    }
}
